requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
paho-mqtt==1.6.1
gunicorn==20.1.0
flask-restful==0.3.9
pymongo==4.5.0
python-dotenv==1.0.0
